Move HTTPError out of internal directory
